Additions & Fixes
Added new Vanilla Legendary Joker replacements: ThorW (Yorick) & ReimMomo (Chicot) - Graphics & name replacements.

Fixed extant issue with the deck skin, the default deck skin should now be selectable in Customise Deck and act as expected. There is no need to switch to High Contrast as, it being the only colour of this skin, it's the default.

Added a config setting to facilitate the future addition of extra challenges. This has no function now, but is a surprise tool that will help us later.

Minor rewording and changes to Joker description changes.

Minor clarification in some comments