New directories
When looking through the repo, I realised that a directory or two was missing. At this point I realised that the repo does not reproduce directories that are empty, so for now I have created some files in the empty directories that I intend on creating content for in the future.