Minor Removals & Changes
Removed currently unused lua files in scripts/retextures as for now, these replacements are done via Malverk, unsure why I didn't do this when I realised that Malverk might not play nice with trying to split them up? I remember making them at the start, thinking I was gonna split everything out to give the player more choice, but I guess I just forgot to remove the files on realisation.

This may change in the future.

Graphic assets moved into their own directories within the assets folder based on whether they are replacing a (or multiple) vanilla asset(s) or are an addition by the mod, and adjusted the relevant code to account for this. This will make looking through the directory in the future a bit easier.

For now, commented out code in the patcher that according to lovely, has no effect.