Major - Additions, Fixes
Gave new Cirno Legendary its own home file and refactored it into the custom joker pipline - Added new NTF Legendary and added iterative joker loader functionality to the main lua.

Re-ordered rename load orders.

Shop flavour text now randomises from a pool defined at the bottom of Misc Renames. The randomise function is called once when starting a run (Because that function is tied to the function loading a saved run calls) and also when defeating a blind (If there was a shop phase-related function that was called as the shop phase starts, that'd be nice. But I can't find it.).

The primary function that drove the functionality of the joker stencils challenge is now defined in the joker stencils challenge lua instead. The pipeline's a little funky, but it should work. Resolved a bug and an exploit (partially, somewhat) with the challenge. Commented out any remaining print debug calls from testing for the time being.

Minor streamlining here and there, w/ alterations to renames as well as the readme.

Known issues: Renaming the Stone Card is kinda funky, mostly doesn't work? This does not appear to be caused by anything to do with this mod and is likely an issue with the newer changes to Steamodded - Waiting on fixes/further investigation.