Major - Additions, Fixes (Yes, again)
Added retextures for Cavendish, Burnt Joker, Seltzer, Hanging Chad & Certificate, with some renames and description flavour text for the last 4.

Minor refactor of Malverk code along with the introduction of a 'Mature References' setting which is disabled by default - This setting informs how the texture pack will populate as well as potential descriptions and renames, etc. Hopefully everything has been made easier to understand.

Additions to the README both to reflect this, as well as some other minor stuff.

Co-Authored-By: DaemonTsun <96687758+DaemonTsun@users.noreply.github.com>