Art credits tooltips, cirMini & Blind Chip fix
Got the whole art credit tooltips system working. - I hate the final structure, but this is the way we have to do things Optionally disableable in the mod config menu.

Added cirMini in place of Wee Joker.

Fixed an unintentional visual issue with Airfryerge.

Updated README in accordance with this and other recent changes.