Major !! Most Important Update Yet !!
The most important part is that during stream, Cirno said yet again that he wouldn't play with the Jimbo's Big Naturals mod - So naturally in response, I made Hannah's boobs bigger. ...That's it, that's the important part.

Obviously this gets altered with the Mature References toggle.

Added retextures for Acrobat, The Idol, Square Joker, Mail-In Rebate & The Sun. Changed Burglar's retexture. Updated metadata accordingly.

Change the palette of Tarot cards, as well as planets and added retextures for the other 11 planets - And an optional, (partially functional) comprehensive renaming of them from Planet Cards to Hu Cards, as well as Celestial Packs to Gensokyo Packs. It's not fully complete, aim to complete it in the coming days. But Planets will basically be done. Disabling the Planets Are Hus will basically revert everything but Neptune & Black Hole (but will revert the planet change). At least, that should be how it works.

Minor alterations to how grabbing art credits on vanilla assets work so that you can't see them on undiscovered or locked Jokers, for example - As well as ensuring that art credits on custom Jokers will actually disable, should the art credits option be disabled.

Other minor tidbits including the README.

Co-Authored-By: DaemonTsun <96687758+DaemonTsun@users.noreply.github.com>