Playing Cards Skin Fix
Fixed an issue where, because the custom deck skin replaced ONLY the face cards without otherwise specifying the rest of the cards or trying to make them also the high contrast versions, it would make it so that only the face cards in that suit were high-contrast, with the rest remaining low-contrast.

For some reason, Steamodded's 'hc_default' property to make it specifically swap the cards that don't get replaced with their high contrast versions doesn't seem to be functional - And other similar fixes I tried to employ either didn't work or caused crashes.

So I ended up using the property for specifically displaying only certain ranks in the 'Customize Deck' screen - Now the skin replaces ALL the cards in the suit, making that entire suit the high-contrast version as intended, while only showing the face cards when selecting it in the 'Customize Deck' screen.

This is probably not the intended way of doing this, but it'll do fine for the time being.