Major - Title Vortex Colour Changes Work + More
FINALLY got the title screen vortex colour changing functionality working (Altered its related config option accordingly as it's functional now - It also defaults to enabled).

Removed all related commented-out code that was the prior attempt at implementing this functionality.

Additionally, all references to Hus (Planets) should now have the appropriate colour. Changed Tarot colour similarly, to match the new palette.

Added retextures for all Gensokyo (Celestial) Booster packs.

Adjusted art credit functionality to accommodate for if we ever include alternate replacements depending on certain conditions (i.e. Whether Planets are Hus is on/off or if we do something different for a normally mature reference item - It's open-ended).

Adjusted The Fool and High Priestess description replacements to properly capture if Planet cards are Hu cards.

(Hopefully) fixed an issue with all custom Legendary Jokers, in which they show as undiscovered with the standard undiscovered Joker description, instead of how vanilla Legendary Jokers appear pre-discovery, which is locked and with the text "Find this Joker from the Soul card." This fix will not be visible to us (Because the keys for these Jokers are already set in our Balatro profiles as discovered) unless we check them from a new profile or change the Jokers' keys - But it should be appearing as intended to anyone installing the mod fresh.

Updates, adjustments & corrections to the README.