More Additions & Changes
Added a mod icon (As it appears in the Steamodded mod menu).

Added retextures for Hallucination, The Family, as well as a placeholder for Cartomancer.

Spilled ketchup all over Girl_DM_'s card :(

Little bit more fleshing out of custom rare graphics.

Learned a bunch about working with wacky stuff like Joker descriptions and trying to change them on the fly and... Hhhhhhhhhhhhh, this is incredibly dumb. All of that to work out a gimmick for The Family, as well as figure some stuff out for future content. Notes were taken, many frustrations were had.

README updated accordingly.

Co-Authored-By: DaemonTsun <96687758+DaemonTsun@users.noreply.github.com>