Near-Total Rewrite of Changing Object Text
A near-total migration of the methods used to rename and change the description of things like Jokers, Tarots, etc. from Steamodded methods to Malverk methods - Near-total because there are some minor hiccups & problem areas I'm not comfortable with migrating all the way over at this time, but most of what you'd expect is changed.

The 'how' of how we make changes to our changes is largely unchanged - The big thing here is that now stuff like name & description changes are more closely tied to texture pack settings in Malverk being enabled, instead of being separate.

This also means that, the way this is written, all the related options in the config menu currently have no effect if the Malverk setting is disabled. Future implementation may see about having it switch between Malverk and Steamodded solutions, but that's a long way down the road. I just wanted to make the switch.

No new assets or major asset changes.