Additions & Changes
Added reskins for Rough Gem, Turtle Bean & Erosion.

More tinkering with how some stuff works under the curtain - Mostly just making a bunch of custom badge definitions into a single, editable function and having all things that use that reference that instead of making the thing each time (I tried making a variable that holds the badge UI code. It either crashes (because UI code seems to destroy the contents of the variable?) or in the case when I try to populate the badge with copy_table(), causes stack overflow.)