v0.0.904, Minor
Adjusted an asset to prepare it for future expansion.

Adjusted Perfectionism.

Added expanded Perfectionism functionality to all Upgraded Legendaries and a few other Jokers.

Added an Upgraded Uncommon.

Fixed an issue with one of the Upgraded Legendaries, additionally fixed a similar issue with Observatory's tooltip, as well as an issue with a Custom Rare that caused crashes. Fixed the sprite behaviour for that Joker's effect.